packages/generator/pyproject.toml
[tool.poetry]
name = "generator"
version = "0.1.0"
description = ""
authors = ["Samuel Duhaime-Morissette <sduhaimemorissette@gmail.com>"]
readme = "README.md"
packages = [{include = "scripts", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.0"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cryptography = "41.0.7"
et-xmlfile = "1.1.0"
idna = "3.6"
msal = "1.26.0"
office365-rest-python-client = "2.5.2"
openpyxl = "3.1.2"
pycparser = "2.21"
pyjwt = "2.8.0"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
requests = "2.31.0"
ruamel-yaml = "0.18.5"
ruamel-yaml-clib = "0.2.8"
typing-extensions = "4.8.0"
urllib3 = "2.1.0"

[tool.poetry.scripts]
generateSurvey = "scripts.generate_survey:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
